May 31 morning: This has two actual bad returns from the AI
The log 07-05 has more information. There are no system failures, but the answers were incorrect for one of them. Does the whole run of 23 questions and then the Kritika 11 questions  from the other dataset. It is "clean" in the sense that only AI events are being logged (service.name=bp-runner) which stands for business process runner